+++ Vidracaria_Portal/Controllers/ClientesController.cs
@@ -45,7 +45,7 @@ namespace Vidracaria_Portal.Controllers
         }
 
         // GET: Clientes/Details/5
-        public async Task<IActionResult> Details(decimal? id)
+        public async Task<IActionResult> Details(int? id)
         {
             if (id == null)
             {
@@ -85,7 +85,7 @@ namespace Vidracaria_Portal.Controllers
         }
 
         // GET: Clientes/Edit/5
-        public async Task<IActionResult> Edit(decimal? id)
+        public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
             {
@@ -105,7 +105,7 @@ namespace Vidracaria_Portal.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(decimal id, [Bind("CodigoCliente,NomeCliente,Email,CidadeCliente,Bairro,NumeroCasa,EstadoCliente,Celular,Documento,DataCadastro,Imagem")] ClientesModel clientesModel)
+        public async Task<IActionResult> Edit(int id, [Bind("CodigoCliente,NomeCliente,Email,CidadeCliente,Bairro,NumeroCasa,EstadoCliente,Celular,Documento,DataCadastro,Imagem")] ClientesModel clientesModel)
         {
             if (id != clientesModel.CodigoCliente)
             {
@@ -136,7 +136,7 @@ namespace Vidracaria_Portal.Controllers
         }
 
         // GET: Clientes/Delete/5
-        public async Task<IActionResult> Delete(decimal? id)
+        public async Task<IActionResult> Delete(int? id)
         {
             if (id == null)
             {
@@ -156,7 +156,7 @@ namespace Vidracaria_Portal.Controllers
         // POST: Clientes/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(decimal id)
+        public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var clientesModel = await _context.Clientes.FindAsync(id);
             _context.Clientes.Remove(clientesModel);
@@ -164,7 +164,7 @@ namespace Vidracaria_Portal.Controllers
             return RedirectToAction(nameof(Index));
         }
 
-        private bool ClientesModelExists(decimal id)
+        private bool ClientesModelExists(int id)
         {
             return _context.Clientes.Any(e => e.CodigoCliente == id);
         }
