+++ Vidracaria_Portal/Controllers/OrcamentosController.cs
@@ -48,7 +48,7 @@ namespace Vidracaria_Portal.Controllers
         }
 
         // GET: Orcamentos/Details/5
-        public async Task<IActionResult> Details(decimal? id)
+        public async Task<IActionResult> Details(int? id)
         {
             if (id == null)
             {
@@ -111,7 +111,7 @@ namespace Vidracaria_Portal.Controllers
         }
 
         // GET: Orcamentos/Edit/5
-        public async Task<IActionResult> Edit(decimal? id)
+        public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
             {
@@ -133,7 +133,7 @@ namespace Vidracaria_Portal.Controllers
         // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(decimal id, [Bind("CodigoOrcamento,Nome,Rua,NumeroCasa,Cidade,Bairro,Celular,Telefone,Estado,TipoId,DataCadastro,Imagem")] OrcamentosModel orcamentosModel, IFormFile NovaImagem)
+        public async Task<IActionResult> Edit(int id, [Bind("CodigoOrcamento,Nome,Rua,NumeroCasa,Cidade,Bairro,Celular,Telefone,Estado,TipoId,DataCadastro,Imagem")] OrcamentosModel orcamentosModel, IFormFile NovaImagem)
         {
             if (id != orcamentosModel.CodigoOrcamento)
             {
@@ -186,7 +186,7 @@ namespace Vidracaria_Portal.Controllers
         }
 
         // GET: Orcamentos/Delete/5
-        public async Task<IActionResult> Delete(decimal? id)
+        public async Task<IActionResult> Delete(int? id)
         {
             if (id == null)
             {
@@ -206,7 +206,7 @@ namespace Vidracaria_Portal.Controllers
         // POST: Orcamentos/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(decimal id)
+        public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var orcamentosModel = await _context.Orcamentos.FindAsync(id);
             _context.Orcamentos.Remove(orcamentosModel);
@@ -214,7 +214,7 @@ namespace Vidracaria_Portal.Controllers
             return RedirectToAction(nameof(Index));
         }
 
-        private bool OrcamentosModelExists(decimal id)
+        private bool OrcamentosModelExists(int id)
         {
             return _context.Orcamentos.Any(e => e.CodigoOrcamento == id);
         }
